Create Global vs Local Variables
Change line 4 to use the var keyword. Notice that the value of my_number in the function is now 14 and outside the function is 7.

Using my_number without the var keyword refers to the global variable that has already been declared outside the function in line 1. However, if you use the var keyword inside a function, it declares a new local variable that only exists within that function.